pyproject.toml
[project]
name = "Sora2watermarkRemover"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiofiles>=24.1.0",
    "diffusers>=0.35.1",
    "einops>=0.8.1",
    "fastapi==0.108.0",
    "ffmpeg-python>=0.2.0",
    "fire>=0.7.1",
    "huggingface-hub>=0.35.3",
    "loguru>=0.7.3",
    "omegaconf>=2.3.0",
    "opencv-python>=4.12.0.88",
    "pydantic>=2.11.10",
    "PyQt5>=5.15.0",
    "requests>=2.32.5",
    "streamlit>=1.50.0",
    "torch==2.9.0+cu130",  # pip3 install torch torchvision --index-url https://download.pytorch.org/whl/cu130
    "torchvision==0.24.0+cu130",  
    "tqdm>=4.67.1",
    "transformers>=4.57.0",
    "uuid>=1.30",
    "uvicorn>=0.35.0",
]

[tool.setuptools]
packages = ["sora2wm"]